Agents move, eat and die

+++ src/Pheromone.java
@@ -1,18 +1,18 @@
 public class Pheromone extends Element {
-    private int life_span;
-    private int radius_effect;
+    private double life_span;
+    private double radius_effect;
 
-    public Pheromone(int x, int y, int life_span, int radius_effect) {
+    public Pheromone(double x, double y, double life_span, double radius_effect) {
         super(x, y);
         this.life_span = life_span;
         this.radius_effect = radius_effect;
     }
 
-    public int getLifeSpan() {
+    public double getLifeSpan() {
         return this.life_span;
     }
 
-    public int getRadiusEffect() {
+    public double getRadiusEffect() {
         return this.radius_effect;
     }
 }
