+++ src/Predator.java
@@ -1,5 +1,5 @@
 public class Predator extends Creature{
-    public Predator(int x, int y, int energyLevel, String classType, double movingSpeed) {
+    public Predator(double x, double y, double energyLevel, String classType, double movingSpeed) {
         super(x, y, energyLevel, classType, movingSpeed);
     }
 }
